+++ AWSLex/AWSLexService.m
@@ -27,7 +27,7 @@
 #import "AWSLexSignature.h"
 
 static NSString *const AWSInfoLex = @"Lex";
-NSString *const AWSLexSDKVersion = @"2.26.6";
+NSString *const AWSLexSDKVersion = @"2.26.7";
 
 
 @interface AWSLexResponseSerializer : AWSJSONResponseSerializer