+++ CHANGELOG.md
@@ -13,6 +13,12 @@
   - Reset completion sources in AWSMobileClient (See [PR #3732](https://github.com/aws-amplify/aws-sdk-ios/pull/3732))
   - Crash fix in AWSMobileClient for call to leave DispatchGroup (See [PR #3951](https://github.com/aws-amplify/aws-sdk-ios/pull/3951))
 
+### Misc. Updates
+- Model updates for the following services
+  - AWSSTS
+  - AWSCloudWatchLogs
+  - AWSChimeSDKMessaging
+
 ## 2.26.6
 
 ### Bug Fixes